requirements.txt
streamlit>=1.28
anthropic>=0.39
python-dotenv>=1.0
scikit-learn>=1.3
pytest>=7.4